+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,9 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <thermal_printer/thermal_printer_plugin.h>
+#include <flutter_thermal_printer/flutter_thermal_printer_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  ThermalPrinterPluginRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("ThermalPrinterPlugin"));
+  FlutterThermalPrinterPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterThermalPrinterPluginCApi"));
 }